Move bloc files into own directory
