+++ integration_test/app_test.dart
@@ -9,8 +9,8 @@ import 'package:app/data/repositories/tasks_repository.dart';
 import 'package:app/data/sources/local/tasks_storage.dart';
 import 'package:app/data/sources/remote/tasks_api.dart';
 import 'package:app/domain/bloc/bloc_dispatcher.dart';
-import 'package:app/domain/bloc/sync_bloc.dart';
-import 'package:app/domain/bloc/tasks_cubit.dart';
+import 'package:app/domain/bloc/sync_bloc/sync_bloc.dart';
+import 'package:app/domain/bloc/tasks_cubit/tasks_cubit.dart';
 import 'package:app/domain/entities/task.dart';
 import 'package:app/main.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
