+++ lib/core/di/dependency_container.dart
@@ -11,8 +11,8 @@ import 'package:app/data/repositories/tasks_repository.dart';
 import 'package:app/data/sources/local/tasks_storage.dart';
 import 'package:app/data/sources/remote/tasks_api.dart';
 import 'package:app/domain/bloc/bloc_dispatcher.dart';
-import 'package:app/domain/bloc/sync_bloc.dart';
-import 'package:app/domain/bloc/tasks_cubit.dart';
+import 'package:app/domain/bloc/sync_bloc/sync_bloc.dart';
+import 'package:app/domain/bloc/tasks_cubit/tasks_cubit.dart';
 import 'package:app/firebase_options.dart';
 import 'package:dio/dio.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
