+++ lib/domain/bloc/sync_bloc/sync_state.dart
