+++ lib/presentation/screens/home_screen/home_screen.dart
@@ -3,8 +3,8 @@ import 'package:app/core/di/dependency_container.dart';
 import 'package:app/core/services/device_info_service.dart';
 import 'package:app/core/services/network_status.dart';
 import 'package:app/domain/bloc/bloc_dispatcher.dart';
-import 'package:app/domain/bloc/sync_bloc.dart';
-import 'package:app/domain/bloc/tasks_cubit.dart';
+import 'package:app/domain/bloc/sync_bloc/sync_bloc.dart';
+import 'package:app/domain/bloc/tasks_cubit/tasks_cubit.dart';
 import 'package:app/core/extensions/l10n.dart';
 import 'package:app/presentation/screens/home_screen/widgets/tasks_visibility_button.dart';
 import 'package:app/presentation/screens/home_screen/widgets/header.dart';
