+++ test/domain/bloc/tasks_cubit_test.dart
@@ -1,5 +1,5 @@
 import 'package:app/core/services/settings_storage.dart';
-import 'package:app/domain/bloc/tasks_cubit.dart';
+import 'package:app/domain/bloc/tasks_cubit/tasks_cubit.dart';
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
